ac_snowman 50%

+++ include/libc64/math64.h
@@ -18,7 +18,7 @@ extern "C" {
 s16 sins(u16);
 s16 coss(u16);
 f32 fatan2(f32, f32);
-f64 fsqrt(f32);
+f32 fsqrt(f32);
 f32 facos(f32);
 
 #ifdef __cplusplus
